change(api): rename BoxStockData to StockData
